requirements.txt

torch>=2.1
torch_geometric>=2.5.2
numpy>=1.23
tqdm>=4.66
pandas>=2.0
ogb>=1.3.6
wandb>=0.17
pyyaml>=6.0.1
